backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-socketio==5.13.0
